\name{miscellaneous wiener model functions}
\alias{is.wiener}
\title{Wiener functions}
\description{
  \code{is.wiener} bla.
}
\usage{
  is.wiener(data)
}
\arguments{
  \item{data}{dataframe with data. Needs a reaction time column and a
  accuracy/response column.}  
}
%\details{}
%\references{} 
%\note{}
%\author{}
\examples{
}
\keyword{wiener}
