\name{wiener diffusion model tests}
\alias{wlrt}
\title{Wiener Diffusion model test functions}
\description{
  \code{wlrt} Likelihood Ratio Test.
}
\usage{
  wlrt(wdmspecific, wdmgeneral)
}
\arguments{
  \item{wdmspecific}{object 1}
  \item{wdmgeneral}{object 2}
}
\details{
  The \code{wlrt} function calls ... }
%\references{} 
%\note{}
%\author{}
\examples{
# todo
}
\keyword{wlrt}
